Restructure Model namespaces; add Model.Draw and shift some util classes to TrialManager.Utils
