+++ TrialManager/Model/Draw/PreferredDayEqualityComparer.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using TrialManager.Model.TrialistDb;
 
-namespace TrialManager.Model
+namespace TrialManager.Model.Draw
 {
     internal class PreferredDayEqualityComparer : IEqualityComparer<Trialist>
     {
