+++ TrialManager/Services/DataExportService.cs
@@ -1,11 +1,10 @@
 ﻿using CsvHelper;
-using System;
 using System.Collections.Generic;
 using System.Globalization;
 using System.IO;
 using System.Threading.Tasks;
-using TrialManager.Model;
 using TrialManager.Model.Csv;
+using TrialManager.Model.Draw;
 
 namespace TrialManager.Services
 {
