+++ TrialManager/Utils/StringSimilarityTool.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.Text.RegularExpressions;
 
-namespace TrialManager.Model
+namespace TrialManager.Utils
 {
     /// <summary>
     /// This class implements string comparison algorithm
