+++ TrialManager/Views/DrawDisplayView.xaml
@@ -6,7 +6,7 @@
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
              xmlns:md="http://materialdesigninxaml.net/winfx/xaml/themes"
              xmlns:mde="clr-namespace:MaterialDesignExtensions.Controls;assembly=MaterialDesignExtensions"
-             xmlns:model="clr-namespace:TrialManager.Model"
+             xmlns:model="clr-namespace:TrialManager.Model.Draw"
              xmlns:s="https://github.com/canton7/Stylet"
              xmlns:sys="clr-namespace:System;assembly=System.Runtime"
              d:DesignHeight="450"
